rename downSwoopRight to just downSwoop
